apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deep-agent-api"
version = "0.1.0"
description = "FastAPI skeleton service for the Deep Agent RAG Stack."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0,<2.0.0",
  "celery[redis]>=5.4.0,<6.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "minio>=7.2.12,<8.0.0",
  "PyJWT[crypto]>=2.10.0,<3.0.0",
  "psycopg2-binary>=2.9.10,<3.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
